// packages/frontend/src/generated/types/ProductResponse.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

/**
 * @description Response DTO for the `product` entity.
 */
export interface ProductResponse {
  /**
   * @type string,null
   */
  addIdentification?: string | null
  /**
   * @type string
   */
  commonName: string
  /**
   * @type string
   */
  createdAt: string
  /**
   * @type string, uuid
   */
  createdBy: string
  /**
   * @type string,null
   */
  deletedAt?: string | null
  /**
   * @type string,null, uuid
   */
  deletedBy?: string | null
  /**
   * @type string, uuid
   */
  id: string
  /**
   * @type boolean
   */
  isComponent: boolean
  /**
   * @type string,null
   */
  longName?: string | null
  /**
   * @type string,null, uuid
   */
  manufacturerId?: string | null
  /**
   * @type string, uuid
   */
  originDbId: string
  /**
   * @type string, uuid
   */
  productGroupId: string
  /**
   * @type string
   */
  updatedAt: string
  /**
   * @type string, uuid
   */
  updatedBy: string
}